+++ crates/noseyparker-cli/src/cmd_scan.rs
@@ -624,7 +624,7 @@ pub fn run(global_args: &args::GlobalArgs, args: &args::ScanArgs) -> Result<()> 
             let matches_summary = datastore.summarize()?;
             let matches_table = crate::cmd_summarize::summary_table(&matches_summary);
             println!();
-            matches_table.print_tty(global_args.use_color())?;
+            matches_table.print_tty(global_args.use_color(std::io::stdout()))?;
         }
 
         println!("\nRun the `report` command next to show finding details.");